--- backend/src/main/java/com/team01/backend/domain/post/controller/ApiV1PostController.java
@@ -1,67 +1,0 @@
-package com.team01.backend.domain.post.controller;
-
-
-import com.team01.backend.domain.post.dto.PostDto;
-import com.team01.backend.domain.post.entity.Post;
-import com.team01.backend.domain.post.service.PostService;
-import com.team01.backend.global.response.ApiResponse;
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("api/v1/posts")
-public class ApiV1PostController {
-
-    private final PostService postService;
-    //private final Rq rq;
-
-    record PostWriteReqBody(
-            @Size(min = 2, message = "제목은 2자 이상이어야 합니다.")
-            @NotBlank(message = "제목은 공백일 수 없습니다.")
-            String title,
-
-            @Size(min = 2, message = "내용은 2자 이상이어야 합니다.")
-            @NotBlank(message = "내용은 공백일 수 없습니다.")
-            String content
-    ){
-    }
-
-    record PostWriteResBody(
-            PostDto postDto,
-            long postsCount
-    ) {
-    }
-
-    // 글 작성 api
-    @PostMapping
-    public ResponseEntity<ApiResponse<PostWriteResBody>> write(
-            @RequestBody @Valid PostWriteReqBody reqBody
-    ) {
-        // User actor = rq.getActor();
-
-
-
-        //Post post = postService.write(actor, reqBody.title, reqBody.content);
-        Post post = postService.write(reqBody.title, reqBody.content);
-
-        long postsCount = postService.count();
-
-        return ResponseEntity.ok(
-                ApiResponse.ofSuccess(
-                        new PostWriteResBody(
-                                new PostDto(post),
-                                postsCount
-                        )
-                )
-        );
-    }
-
-}
