+++ backend/src/main/java/com/team01/backend/domain/post/service/PostService.java
@@ -5,11 +5,16 @@ import com.team01.backend.domain.post.entity.Post;
 import com.team01.backend.domain.post.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 
 import java.util.List;
+import java.util.Optional;
+
 
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class PostService {
 
     private final PostRepository postRepository;
@@ -20,7 +25,7 @@ public class PostService {
 //        return postRepository.save(post);
 //    }
 
-
+    @Transactional
     public Post write(String title, String content) {
         Post post = new Post(title, content);
         return postRepository.save(post);
@@ -35,6 +40,17 @@ public class PostService {
                 .stream()
                 .map(PostResponseDto::new)
                 .toList();
+    }
+
+    public Optional<Post> findById(Long id) {return postRepository.findById(id);}
+
+    @Transactional
+    public Post modify(Long id, String title, String content) {
+        Post post = postRepository.findById(id).get();
+        post.update(title, content);
+
+        return post;
+
     }
 
 }
