--- backend/src/test/java/com/team01/backend/domain/post/controller/ApiV1PostControllerTest.java
@@ -1,149 +1,0 @@
-package com.team01.backend.domain.post.controller;
-
-import com.team01.backend.domain.post.repository.PostRepository;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.ResultActions;
-import org.springframework.transaction.annotation.Transactional;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-
-@SpringBootTest
-@ActiveProfiles("test")
-@AutoConfigureMockMvc
-@Transactional
-public class ApiV1PostControllerTest {
-
-    @Autowired
-    private MockMvc mvc;
-
-    @Autowired
-    private PostRepository postRepository;
-
-    @Test
-    @DisplayName("글 작성")
-    void t1() throws Exception {
-        String title = "제목입니다.";
-        String content = "내용입니다.";
-
-        ResultActions resultActions = mvc
-                .perform(
-                        post("/api/v1/posts")
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content("""
-                                        {
-                                            "title": "%s",
-                                            "content": "%s"
-                                        }
-                                        """.formatted(title, content))
-                )
-                .andDo(print());
-
-        resultActions
-                .andExpect(handler().handlerType(ApiV1PostController.class))
-                .andExpect(handler().methodName("write"))
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.success").value(true))
-                .andExpect(jsonPath("$.data.postDto.id").exists())
-                .andExpect(jsonPath("$.data.postDto.title").value(title))
-                .andExpect(jsonPath("$.data.postDto.content").value(content))
-                .andExpect(jsonPath("$.data.postDto.createdAt").exists())
-                .andExpect(jsonPath("$.data.postDto.modifiedAt").exists());
-    }
-
-    @Test
-    @DisplayName("글 작성 실패 - 제목이 입력되지 않은 경우")
-    void t2() throws Exception {
-
-        String title = "";
-        String content = "내용입니다.";
-
-
-        ResultActions resultActions = mvc
-                .perform(
-                        post("/api/v1/post")
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content("""
-                                    {
-                                        "title": "%s",
-                                        "content": "%s"
-                                    }
-                                    """.formatted(title, content))
-                )
-                .andDo(print());
-
-        resultActions
-                .andExpect(handler().handlerType(ApiV1PostController.class))
-                .andExpect(handler().methodName("write"))
-                .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("$.success").value(false))
-                .andExpect(jsonPath("$.code").value("INVALID_INPUT"))
-                .andExpect(jsonPath("$.message").exists());
-    }
-
-    @Test
-    @DisplayName("글 작성, 내용이 입력되지 않은 경우")
-    void t3() throws Exception {
-        String title = "제목입니다.";
-        String content = "";
-
-        ResultActions resultActions = mvc
-                .perform(
-                        post("/api/v1/post")
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content(
-                                        """
-                                        {
-                                            "title": "%s",
-                                            "content": "%s"
-                                        }
-                                        """.formatted(title, content))
-                )
-                .andDo(print());
-
-        resultActions
-                .andExpect(handler().handlerType(ApiV1PostController.class))
-                .andExpect(handler().methodName("write"))
-                .andExpect(jsonPath("$.success").value(false))
-                .andExpect(jsonPath("$.code").value("INVALID_INPUT"))
-                .andExpect(jsonPath("$.message").exists());
-    }
-
-    @Test
-    @DisplayName("글 작성 실패 - JSON 양식이 잘못된 경우")
-    void t4() throws Exception {
-
-        String title = "제목입니다.";
-        String content = "내용입니다.";
-
-        ResultActions resultActions = mvc
-                .perform(
-                        post("/api/v1/post")
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .content("""
-                                    {
-                                        "title": "%s"   // , 누락
-                                        "content": "%s"
-                                    }
-                                    """.formatted(title, content))
-                )
-                .andDo(print());
-
-        resultActions
-                .andExpect(handler().handlerType(ApiV1PostController.class))
-                .andExpect(handler().methodName("write"))
-                .andExpect(status().isBadRequest())
-                .andExpect(jsonPath("$.success").value(false))
-                .andExpect(jsonPath("$.code").value("INVALID_JSON"))
-                .andExpect(jsonPath("$.message").value("잘못된 형식의 JSON 데이터입니다."));
-    }
-}
